chore(components): migrate aiplatform.automl_training_job components to v1.automl.training_job
PiperOrigin-RevId: 526719420
