+++ components/google-cloud/google_cloud_pipeline_components/v1/automl/training_job/automl_image_training_job/component.yaml
