+++ components/google-cloud/tests/v1/testdata/automl_forecasting_pipeline.json
