+++ components/google-cloud/tests/v1/testdata/automl_image_pipeline.json
