+++ components/google-cloud/tests/v1/testdata/automl_tabular_pipeline.json
