+++ components/google-cloud/tests/v1/testdata/automl_text_pipeline.json
