+++ components/google-cloud/tests/v1/testdata/automl_video_pipeline.json
